xrstring_manager.cpp is created and the content is added, xrstring.cpp unnecessary content is removed
