+++ src/xrCore/CMakeLists.txt
@@ -386,6 +386,8 @@ target_sources_grouped(
     xrsharedmem.h
     xrstring.cpp
     xrstring.h
+    xrstring_manager.cpp
+    xrstring_manager.h
 )
 
 target_sources_grouped(
